# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Tamiyo API
  description: >
    REST API for managing a Magic: The Gathering card collection —
    cards, physical storage (binders, boxes, deckboxes), and decks.
    Every endpoint except /health and /auth/register, /auth/login,
    /auth/refresh, /auth/logout requires a Bearer access token
    (obtained from /auth/register or /auth/login) and scopes all data
    to the authenticated user.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /deck:
    get:
      tags: [decks]
      summary: List decks
      description: List the authenticated user's decks, with optional filtering and pagination.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
          description: >
            Only return decks whose format exactly matches (case-insensitive).
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-based page number.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          description: Number of decks per page (max 100).
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, -name, added, -added, updated, -updated]
            default: -updated
          description: >
            Sort field, with an optional "-" prefix for descending order.
            id is always used as a secondary tie-breaker for a stable order.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDecks"
        "400":
          description: page or limit is invalid, or sort is not one of the allowed values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [decks]
      summary: Create a deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeckRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "400":
          description: >
            Missing required field, or commander_id does not reference
            an existing card owned by the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /deck/{id}:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    get:
      tags: [decks]
      summary: Get a deck by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "400":
          $ref: "#/components/responses/InvalidID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Deck not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [decks]
      summary: Partially update a deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeckRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "400":
          description: Invalid id, invalid body, or invalid commander_id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Deck not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [decks]
      summary: Delete a deck
      description: >
        Removes all card_deck associations for this deck. Cards themselves
        are never deleted.
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Deck not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /deck/{id}/cards:
    parameters:
      - $ref: "#/components/parameters/DeckID"
    get:
      tags: [deck-cards]
      summary: List the cards in a deck
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, -name, added, -added, updated, -updated, mana_value, -mana_value]
            default: -updated
          description: >
            Sort field, with an optional "-" prefix for descending order.
            id is always used as a secondary tie-breaker for a stable order.
      responses:
        "200":
          description: OK (empty array if the deck has no cards)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/InvalidID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Deck not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /deck/{id}/cards/{card_id}:
    parameters:
      - $ref: "#/components/parameters/DeckID"
      - $ref: "#/components/parameters/CardIDPath"
    put:
      tags: [deck-cards]
      summary: Add a card to a deck
      description: >
        Idempotent: calling this again with the same id/card_id succeeds
        silently if the card is already in the deck.
      responses:
        "204":
          description: No Content
        "400":
          description: id or card_id is not a valid integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Deck not found or card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [deck-cards]
      summary: Remove a card from a deck
      description: >
        Idempotent: always succeeds, whether or not the card was actually
        linked to the deck, or whether the deck/card exist at all. The
        guaranteed end state is "this card is not in this deck".
      responses:
        "204":
          description: No Content
        "400":
          description: id or card_id is not a valid integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token returned by /auth/register, /auth/login or
        /auth/refresh. Short-lived (JWT_ACCESS_TOKEN_TTL_MINUTES,
        15 minutes by default) — use /auth/refresh to obtain a new one
        without logging in again.
  parameters:
...
    DeckID:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Deck ID
    CardIDPath:
      name: card_id
      in: path
      required: true
      schema:
        type: integer
      description: Card ID
  responses:
    InvalidID:
      description: id is not a valid integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server/database error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Card:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scryfall_id:
          type: string
          format: uuid
        set_code:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
        storage_id:
          type: integer
          nullable: true
        mana_value:
          type: number
          description: Converted mana cost (CMC).
        added:
          type: string
        updated:
          type: string
      required:
        - id
        - name
        - scryfall_id
        - set_code
        - collector_number
        - foil
        - mana_value
        - added
        - updated
...
    Deck:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        format:
          type: string
        commander_id:
          type: integer
          nullable: true
        card_count:
          type: integer
        added:
          type: string
        updated:
          type: string
      required: [id, name, format, card_count, added, updated]
    PaginatedDecks:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Deck"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required: [data, page, limit, total, total_pages]
    CreateDeckRequest:
      type: object
      properties:
        name:
          type: string
        format:
          type: string
        commander_id:
          type: integer
          minimum: 1
      required: [name, format]
    UpdateDeckRequest:
      type: object
      description: All fields optional; only included fields are modified.
      properties:
        name:
          type: string
        format:
          type: string
        commander_id:
          type: integer
          minimum: 1
        clear_commander_id:
          type: boolean
          description: Set to true to explicitly clear commander_id (set it to null).
...